INS-136: Add docs for artifact manager interface.
